Move zipkin config to init
