+++ pkg/standalone/standalone.go
@@ -18,7 +18,6 @@ import (
 	"os/exec"
 	"os/user"
 	"path"
-	"path/filepath"
 	path_filepath "path/filepath"
 	"runtime"
 	"strings"
@@ -41,9 +40,11 @@ const (
 	daprLatestVersion                 = "latest"
 	daprDefaultLinuxAndMacInstallPath = "/usr/local/bin"
 	daprDefaultWindowsInstallPath     = "c:\\dapr"
-	daprDefaultRedisHost              = "localhost"
+	daprDefaultHost                   = "localhost"
 	pubSubYamlFileName                = "pubsub.yaml"
 	stateStoreYamlFileName            = "statestore.yaml"
+	zipkinYamlFileName                = "zipkin.yaml"
+	defaultConfigFileName             = "default.yaml"
 
 	// DaprPlacementContainerName is the container name of placement service
 	DaprPlacementContainerName = "dapr_placement"
@@ -52,6 +53,36 @@ const (
 	// DaprZipkinContainerName is the container name of zipkin
 	DaprZipkinContainerName = "dapr_zipkin"
 )
+
+type configuration struct {
+	APIVersion string `yaml:"apiVersion"`
+	Kind       string `yaml:"kind"`
+	Metadata   struct {
+		Name string `yaml:"name"`
+	} `yaml:"metadata"`
+	Spec struct {
+		Tracing struct {
+			SamplingRate string `yaml:"samplingRate"`
+		} `yaml:"tracing"`
+	} `yaml:"spec"`
+}
+
+type component struct {
+	APIVersion string `yaml:"apiVersion"`
+	Kind       string `yaml:"kind"`
+	Metadata   struct {
+		Name string `yaml:"name"`
+	} `yaml:"metadata"`
+	Spec struct {
+		Type     string                  `yaml:"type"`
+		Metadata []componentMetadataItem `yaml:"metadata"`
+	} `yaml:"spec"`
+}
+
+type componentMetadataItem struct {
+	Name  string `yaml:"name"`
+	Value string `yaml:"value"`
+}
 
 func isInstallationRequired(installLocation, requestedVersion string) bool {
 	var destDir string
@@ -255,7 +286,7 @@ func runZipkin(wg *sync.WaitGroup, errorChan chan<- error, dir, version string, 
 func runRedis(wg *sync.WaitGroup, errorChan chan<- error, dir, version string, dockerNetwork string, installLocation string, redisHost string) {
 	defer wg.Done()
 
-	if redisHost != daprDefaultRedisHost {
+	if redisHost != daprDefaultHost {
 		// A non-default Redis host is specified. No need to start the redis container
 		fmt.Printf("You have specified redis-host: %s. Make sure you have a redis server running there.\n", redisHost)
 		return
@@ -452,11 +483,13 @@ func installDaprBinary(wg *sync.WaitGroup, errorChan chan<- error, dir, version 
 func createComponentsDir(wg *sync.WaitGroup, errorChan chan<- error, dir, version string, dockerNetwork string, installLocation string, redisHost string) {
 	defer wg.Done()
 
+	var err error
+
 	// Make default components directory
 	componentsDir := GetDefaultFolderPath(DefaultComponentsDirName)
-	_, err := os.Stat(componentsDir)
+	_, err = os.Stat(componentsDir)
 	if os.IsNotExist(err) {
-		errDir := os.MkdirAll(componentsDir, 0777)
+		errDir := os.MkdirAll(componentsDir, 0755)
 		if errDir != nil {
 			errorChan <- fmt.Errorf("error creating default components folder: %s", errDir)
 			return
@@ -464,7 +497,7 @@ func createComponentsDir(wg *sync.WaitGroup, errorChan chan<- error, dir, versio
 	}
 
 	// Make default config directory
-	configDir := GetDefaultFolderPath(DefaultComponentsDirName)
+	configDir := GetDefaultFolderPath(defaultConfigDirName)
 	_, err = os.Stat(configDir)
 	if os.IsNotExist(err) {
 		errDir := os.MkdirAll(configDir, 0755)
@@ -475,9 +508,28 @@ func createComponentsDir(wg *sync.WaitGroup, errorChan chan<- error, dir, versio
 	}
 
 	os.Chmod(componentsDir, 0777)
+	os.Chmod(configDir, 0777)
 
-	createRedisPubSub(redisHost, componentsDir)
-	createRedisStateStore(redisHost, componentsDir)
+	err = createRedisPubSub(redisHost, componentsDir)
+	if err != nil {
+		errorChan <- fmt.Errorf("error creating redis pubsub component file: %s", err)
+		return
+	}
+	err = createRedisStateStore(redisHost, componentsDir)
+	if err != nil {
+		errorChan <- fmt.Errorf("error creating redis statestore component file: %s", err)
+		return
+	}
+	err = createZipkinComponent(daprDefaultHost, componentsDir)
+	if err != nil {
+		errorChan <- fmt.Errorf("error creating zipkin component file: %s", err)
+		return
+	}
+	err = createDefaultConfigurtion(configDir)
+	if err != nil {
+		errorChan <- fmt.Errorf("error creating default configuration file: %s", err)
+		return
+	}
 }
 
 func makeExecutable(filepath string) error {
@@ -710,14 +762,10 @@ func createRedisStateStore(redisHost string, componentsPath string) error {
 		return err
 	}
 
-	filePath := filepath.Join(componentsPath, stateStoreYamlFileName)
-	err = ioutil.WriteFile(filePath, b, 0600)
-	if err != nil {
-		return err
-	}
-	os.Chmod(filePath, 0644)
+	filePath := path_filepath.Join(componentsPath, stateStoreYamlFileName)
+	err = checkAndOverWrite(filePath, b)
 
-	return nil
+	return err
 }
 
 func createRedisPubSub(redisHost string, componentsPath string) error {
@@ -744,12 +792,69 @@ func createRedisPubSub(redisHost string, componentsPath string) error {
 		return err
 	}
 
-	filePath := filepath.Join(componentsPath, pubSubYamlFileName)
-	err = ioutil.WriteFile(filePath, b, 0600)
+	filePath := path_filepath.Join(componentsPath, pubSubYamlFileName)
+	err = checkAndOverWrite(filePath, b)
+
+	return err
+}
+
+func createDefaultConfigurtion(configDir string) error {
+	defaultConfig := configuration{
+		APIVersion: "dapr.io/v1alpha1",
+		Kind:       "Configuration",
+	}
+	defaultConfig.Metadata.Name = "daprConfig"
+	defaultConfig.Spec.Tracing.SamplingRate = "1"
+
+	b, err := yaml.Marshal(&defaultConfig)
 	if err != nil {
 		return err
 	}
-	os.Chmod(filePath, 0644)
 
+	filePath := path_filepath.Join(configDir, defaultConfigFileName)
+	err = checkAndOverWrite(filePath, b)
+
+	return err
+}
+
+func createZipkinComponent(zipkinHost string, componentsPath string) error {
+	zipKinComponent := component{
+		APIVersion: "dapr.io/v1alpha1",
+		Kind:       "Component",
+	}
+	zipKinComponent.Metadata.Name = "zipkin"
+	zipKinComponent.Spec.Type = "exporters.zipkin"
+	zipKinComponent.Spec.Metadata = []componentMetadataItem{
+		{
+			Name:  "enabled",
+			Value: "true",
+		},
+		{
+			Name:  "exporterAddress",
+			Value: fmt.Sprintf("http://%s:9411/api/v2/spans", zipkinHost),
+		},
+	}
+
+	b, err := yaml.Marshal(&zipKinComponent)
+	if err != nil {
+		return err
+	}
+
+	filePath := path_filepath.Join(componentsPath, zipkinYamlFileName)
+	err = checkAndOverWrite(filePath, b)
+
+	return err
+}
+
+func checkAndOverWrite(filePath string, b []byte) error {
+	_, err := os.Stat(filePath)
+	if os.IsNotExist(err) {
+		// #nosec G306
+		if err = ioutil.WriteFile(filePath, b, 0644); err != nil {
+			return err
+		}
+	} else {
+		fmt.Printf("file %s exists in the default dapr path", filePath)
+	}
 	return nil
 }
